Support to add unsigned long long and long long as fields
To be able to send large integers such as nanosecond timestamps for
syslog and Chronograf logging from esp32.

Currently casting to a smaller data type results in truncation, double
will not have the i suffix and casting to string (+"i") before calling
addField gets wrapped in quotation marks so database treats it as
string.

Fixes #164